requirements.txt
#PyMySQL==1.1.0
#redis==5.0.1
python-dotenv==1.0.0
#Flask==3.0.0
PyYAML==6.0
black
#flask-restful
#flask-httpauth
pandas
seaborn
pymongo
matplotlib
